requirements.txt
Django==3.1.4
django-crispy-forms==1.10.0
Pillow==8.0.1